EXPW is now a float so we can decrease by non-unity values. this was done so that there does not need to be one event per domain make the event query smarter so it works when there have been no events if there have been no events, default to one day ago as last time
git-svn-id: file:///home/mjibson/backup/arcrepo/trunk@940 157e5a38-50d5-0310-a571-dcc21a78b580

+++ ci4/utility/Event.inc.php
@@ -73,7 +73,7 @@ function expwDecrease($id, $last)
 
 		for($i = 0; $i < count($res); $i++)
 		{
-			$dec = (int)($hours / $res[$i]['domain_expw_time']);
+			$dec = $hours / $res[$i]['domain_expw_time'];
 
 			if($dec)
 			{
